requirements.txt
astra-toolbox
Cython
h5py
imageio
ipython
ipython-genutils
matplotlib
nibabel
numpy
odl
opencv-python
Pillow
pyparsing
python-dateutil
requests
scikit-image
scikit-learn
scipy
SimpleITK
six
tensorboard
tensorboardX
tensorflow
tensorflow-estimator
tensorflow-gpu
tensorflow-tensorboard
torch
torchfile
torchstat
torchsummary
torchvision
tqdm

